Move SystemComponent to own package
